+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **\[Python/C++\]** Add `BoundsTrimFileExample` and `ResolutionTrimFileExample` to example-writer files
 - **\[Python/C++\]** Add `Classification` getter and setter to `Points`
 - **\[Python/C++\]** Add `GetAllChildrenPoints` to `Reader`
+- **\[Python\]** Make `Node` and `VoxelKey` picklable
 
 ### Changed